Fix CI: remove duplicate test files from src/test, fix incubator javadoc
- Delete fileconfig tests from src/test/ (they were moved to testDeclarativeConfig
  in the previous commit but deletions were not staged)
- Fix ViewConfig/ViewConfigCustomizer javadoc: replace broken {@link} cross-module
  reference to DeclarativeConfigurationParser with {@code} (class moved to
  autoconfigure module which is not a main-scope dep of incubator)
- Delete incubator META-INF/services ComponentProvider registration (ServiceResourceDetector
  moved to autoconfigure module)
