--- sdk-extensions/autoconfigure/src/test/java/io/opentelemetry/sdk/extension/incubator/fileconfig/DeclarativeConfigurationBuilderTest.java
@@ -1,116 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.sdk.extension.incubator.fileconfig;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-
-import io.opentelemetry.api.incubator.config.DeclarativeConfigProperties;
-import io.opentelemetry.sdk.logs.export.LogRecordExporter;
-import io.opentelemetry.sdk.metrics.export.MetricExporter;
-import io.opentelemetry.sdk.trace.export.SpanExporter;
-import org.junit.jupiter.api.Test;
-
-class DeclarativeConfigurationBuilderTest {
-
-  @Test
-  void spanExporterCustomizer_Single() {
-    DeclarativeConfigurationBuilder builder = new DeclarativeConfigurationBuilder();
-
-    builder.addSpanExporterCustomizer(
-        SpanExporter.class, (exporter, properties) -> mock(SpanExporter.class));
-
-    assertThat(builder.getSpanExporterCustomizers()).hasSize(1);
-  }
-
-  @Test
-  void spanExporterCustomizer_Multiple_Compose() {
-    DeclarativeConfigurationBuilder builder = new DeclarativeConfigurationBuilder();
-
-    builder.addSpanExporterCustomizer(
-        SpanExporter.class, (exporter, properties) -> mock(SpanExporter.class));
-    builder.addSpanExporterCustomizer(
-        SpanExporter.class, (exporter, properties) -> mock(SpanExporter.class));
-
-    assertThat(builder.getSpanExporterCustomizers()).hasSize(2);
-  }
-
-  @Test
-  void metricExporterCustomizer_Single() {
-    DeclarativeConfigurationBuilder builder = new DeclarativeConfigurationBuilder();
-
-    builder.addMetricExporterCustomizer(
-        MetricExporter.class, (exporter, properties) -> mock(MetricExporter.class));
-
-    assertThat(builder.getMetricExporterCustomizers()).hasSize(1);
-  }
-
-  @Test
-  void metricExporterCustomizer_Multiple_Compose() {
-    DeclarativeConfigurationBuilder builder = new DeclarativeConfigurationBuilder();
-
-    builder.addMetricExporterCustomizer(
-        MetricExporter.class, (exporter, properties) -> mock(MetricExporter.class));
-    builder.addMetricExporterCustomizer(
-        MetricExporter.class, (exporter, properties) -> mock(MetricExporter.class));
-
-    assertThat(builder.getMetricExporterCustomizers()).hasSize(2);
-  }
-
-  @Test
-  void logRecordExporterCustomizer_Single() {
-    DeclarativeConfigurationBuilder builder = new DeclarativeConfigurationBuilder();
-
-    builder.addLogRecordExporterCustomizer(
-        LogRecordExporter.class, (exporter, properties) -> mock(LogRecordExporter.class));
-
-    assertThat(builder.getLogRecordExporterCustomizers()).hasSize(1);
-  }
-
-  @Test
-  void logRecordExporterCustomizer_Multiple_Compose() {
-    DeclarativeConfigurationBuilder builder = new DeclarativeConfigurationBuilder();
-
-    builder.addLogRecordExporterCustomizer(
-        LogRecordExporter.class, (exporter, properties) -> mock(LogRecordExporter.class));
-    builder.addLogRecordExporterCustomizer(
-        LogRecordExporter.class, (exporter, properties) -> mock(LogRecordExporter.class));
-
-    assertThat(builder.getLogRecordExporterCustomizers()).hasSize(2);
-  }
-
-  @Test
-  void customizer_ClosesOriginalWhenReplaced() throws Exception {
-    SpanExporter original = mock(SpanExporter.class);
-    SpanExporter replacement = mock(SpanExporter.class);
-    DeclarativeConfigProperties props = mock(DeclarativeConfigProperties.class);
-
-    DeclarativeConfigurationBuilder.Customizer<SpanExporter> customizer =
-        new DeclarativeConfigurationBuilder.Customizer<>(
-            SpanExporter.class, (exporter, properties) -> replacement);
-
-    SpanExporter result = customizer.maybeCustomize(original, "test", props);
-
-    assertThat(result).isSameAs(replacement);
-    verify(original).close();
-  }
-
-  @Test
-  void customizer_DoesNotCloseWhenSameInstance() throws Exception {
-    SpanExporter exporter = mock(SpanExporter.class);
-    DeclarativeConfigProperties props = mock(DeclarativeConfigProperties.class);
-
-    DeclarativeConfigurationBuilder.Customizer<SpanExporter> customizer =
-        new DeclarativeConfigurationBuilder.Customizer<>(SpanExporter.class, (e, properties) -> e);
-
-    SpanExporter result = customizer.maybeCustomize(exporter, "test", props);
-
-    assertThat(result).isSameAs(exporter);
-    verify(exporter, never()).close();
-  }
-}
