--- sdk-extensions/autoconfigure/src/test/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider
@@ -1,14 +1,0 @@
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.MetricExporterComponentProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.SpanExporterComponentProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.LogRecordExporterComponentProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.TextMapPropagatorComponentProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.SamplerComponentProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.SpanProcessorComponentProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.LogRecordProcessorComponentProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.ResourceComponentProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.ResourceFirstComponentProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.ResourceSecondComponentProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.ContainerResourceProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.HostResourceProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.OsResourceProvider
-io.opentelemetry.sdk.extension.incubator.fileconfig.component.ProcessResourceProvider
