--- sdk-extensions/autoconfigure/src/test/resources/aggregation-args.yaml
@@ -1,21 +1,0 @@
-- selector:
-    instrument_type: HISTOGRAM
-  view:
-    aggregation: explicit_bucket_histogram
-    aggregation_args:
-      bucket_boundaries: [1.0, 2.0, 5.0]
-- selector:
-    instrument_type: HISTOGRAM
-  view:
-    aggregation: explicit_bucket_histogram
-    aggregation_args:
-      bucket_boundaries:
-        - 1.0
-        - 2.0
-        - 5.0
-- selector:
-    instrument_type: HISTOGRAM
-  view:
-    aggregation: exponential_bucket_histogram
-    aggregation_args:
-      max_buckets: 20
