--- sdk-extensions/autoconfigure/src/test/resources/empty-view-config.yaml
@@ -1,8 +1,0 @@
-- selector:
-    instrument_name: name1
-    instrument_type: COUNTER
-    instrument_unit: ms
-    meter_name: meterName1
-    meter_version: 1.0.0
-    meter_schema_url: http://example1.com
-  view:
