--- sdk-extensions/autoconfigure/src/test/resources/full-config.yaml
@@ -1,28 +1,0 @@
-- selector:
-    instrument_name: name1
-    instrument_type: COUNTER
-    instrument_unit: ms
-    meter_name: meterName1
-    meter_version: 1.0.0
-    meter_schema_url: http://example1.com
-  view:
-    name: name1
-    description: description1
-    aggregation: sum
-    attribute_keys:
-      - foo
-      - bar
-- selector:
-    instrument_name: name2
-    instrument_type: COUNTER
-    instrument_unit: s
-    meter_name: meterName2
-    meter_version: 2.0.0
-    meter_schema_url: http://example2.com
-  view:
-    name: name2
-    description: description2
-    aggregation: last_value
-    attribute_keys:
-      - baz
-      - qux
