--- sdk-extensions/incubator/src/main/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider
@@ -1,1 +1,0 @@
-io.opentelemetry.sdk.extension.incubator.fileconfig.ServiceResourceDetector
